plugin crop and rotate
