+++ packages/core/src/index.js
@@ -1054,7 +1054,7 @@ Jimp.loadFont = function(file, cb) {
  * The emitted event comes with a object parameter to the listener with the
  * `methodName` as one attribute.
  */
-function jimpEvMethod(methodName, evName, method) {
+export function jimpEvMethod(methodName, evName, method) {
     const evNameBefore = 'before-' + evName;
     const evNameAfter = evName.replace(/e$/, '') + 'ed';
 
@@ -1125,7 +1125,8 @@ jimpEvMethod('clone', 'clone', function(cb) {
  * @param {string} methodName name of the method
  * @param {function} method to watch changes for
  */
-function jimpEvChange(methodName, method) {
+export function jimpEvChange(methodName, method) {
+    console.log(methodName);
     jimpEvMethod(methodName, 'change', method);
 }
 
@@ -1184,56 +1185,6 @@ jimpEvChange('scan', function(x, y, w, h, f, cb) {
     }
 
     return result;
-});
-
-/**
- * Crops the image at a given point to a give size
- * @param {number} x the x coordinate to crop form
- * @param {number} y the y coordinate to crop form
- * @param w the width of the crop region
- * @param h the height of the crop region
- * @param {function(Error, Jimp)} cb (optional) a callback for when complete
- * @returns {Jimp} this for chaining of methods
- */
-jimpEvChange('crop', function(x, y, w, h, cb) {
-    if (typeof x !== 'number' || typeof y !== 'number')
-        return throwError.call(this, 'x and y must be numbers', cb);
-    if (typeof w !== 'number' || typeof h !== 'number')
-        return throwError.call(this, 'w and h must be numbers', cb);
-
-    // round input
-    x = Math.round(x);
-    y = Math.round(y);
-    w = Math.round(w);
-    h = Math.round(h);
-
-    if (x === 0 && w === this.bitmap.width) {
-        // shortcut
-        const start = (w * y + x) << 2;
-        const end = (start + h * w) << (2 + 1);
-
-        this.bitmap.data = this.bitmap.data.slice(start, end);
-    } else {
-        const bitmap = Buffer.allocUnsafe(w * h * 4);
-        let offset = 0;
-
-        this.scanQuiet(x, y, w, h, function(x, y, idx) {
-            const data = this.bitmap.data.readUInt32BE(idx, true);
-            bitmap.writeUInt32BE(data, offset, true);
-            offset += 4;
-        });
-
-        this.bitmap.data = bitmap;
-    }
-
-    this.bitmap.width = w;
-    this.bitmap.height = h;
-
-    if (isNodePattern(cb)) {
-        cb.call(this, null, this);
-    }
-
-    return this;
 });
 
 if (process.env.ENVIRONMENT === 'BROWSER') {
