+++ packages/jimp/package.json
@@ -64,7 +64,9 @@
     "@jimp/bmp": "0.3.5",
     "@jimp/tiff": "0.3.5",
     "@jimp/plugin-resize": "0.3.5",
-    "@jimp/plugin-blit": "0.3.5"
+    "@jimp/plugin-blit": "0.3.5",
+    "@jimp/plugin-rotate": "0.3.5",
+    "@jimp/plugin-crop": "0.3.5"
   },
   "devDependencies": {
     "babelify": "^9.0.0",
