+++ packages/jimp/src/index.js
@@ -2,5 +2,10 @@ import configure from '@jimp/custom';
 
 export default configure({
     types: ['@jimp/jpeg', '@jimp/png', '@jimp/bmp', '@jimp/tiff', '@jimp/gif'],
-    plugins: ['@jimp/plugin-resize', '@jimp/plugin-blit']
+    plugins: [
+        '@jimp/plugin-resize',
+        '@jimp/plugin-blit',
+        '@jimp/plugin-rotate',
+        '@jimp/plugin-crop'
+    ]
 });
