+++ packages/plugin-crop/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-crop
+
+crop an image.
