+++ packages/plugin-rotate/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-rotate
+
+Rotate an image.
